Finished IDB Memory for WasmBoy
