+++ lib/memory/idb.js
@@ -1,0 +1,58 @@
+// Get our idb instance, and initialize to asn idb-keyval
+// This is so we don't get the default keyval DB name. And will allow
+// Parent projects to use the slimmer idb keyval
+// https://www.npmjs.com/package/idb
+const idb = require('idb');
+
+// Get our idb dPromise
+const dbPromise = idb.open('wasmboy', 1, upgradeDB => {
+  upgradeDB.createObjectStore('keyval');
+});
+
+// Get our idb-keyval instance
+export const idbKeyval = {
+  get(key) {
+    return dbPromise.then(db => {
+      return db.transaction('keyval')
+        .objectStore('keyval').get(key);
+    });
+  },
+  set(key, val) {
+    return dbPromise.then(db => {
+      const tx = db.transaction('keyval', 'readwrite');
+      tx.objectStore('keyval').put(val, key);
+      return tx.complete;
+    });
+  },
+  delete(key) {
+    return dbPromise.then(db => {
+      const tx = db.transaction('keyval', 'readwrite');
+      tx.objectStore('keyval').delete(key);
+      return tx.complete;
+    });
+  },
+  clear() {
+    return dbPromise.then(db => {
+      const tx = db.transaction('keyval', 'readwrite');
+      tx.objectStore('keyval').clear();
+      return tx.complete;
+    });
+  },
+  keys() {
+    return dbPromise.then(db => {
+      const tx = db.transaction('keyval');
+      const keys = [];
+      const store = tx.objectStore('keyval');
+
+      // This would be store.getAllKeys(), but it isn't supported by Edge or Safari.
+      // openKeyCursor isn't supported by Safari, so we fall back
+      (store.iterateKeyCursor || store.iterateCursor).call(store, cursor => {
+        if (!cursor) return;
+        keys.push(cursor.key);
+        cursor.continue();
+      });
+
+      return tx.complete.then(() => keys);
+    });
+  }
+};
