+++ lib/wasmboy.js
@@ -125,6 +125,11 @@ class WasmBoyLib {
     this.paused = true;
   }
 
+  // Function to return the current game object in memory
+  getWasmBoyMemoryForLoadedGame() {
+    return WasmBoyMemory.getCartridgeObject();
+  }
+
   saveState() {
     // Pause the game in case it was running, and set to not ready
     this.pauseGame();
